Rename: dto 디렉토리 위치 변경 (application -> interfaces)
- 컨트롤러에서 사용하는 것이기에 interfaces에 dto가 위치하는 것이 맞다고 판단
