+++ src/point/interfaces/dto/point-response.dto.ts
